Core/Base: Warning fix

+++ include/inviwo/core/datastructures/spatialdata.h
@@ -316,13 +316,13 @@ StructuredGridEntity<N>::StructuredGridEntity(const Vector<N, size_t>& dimension
                                               const Vector<N, float>& spacing)
     : SpatialEntity<N>(), dimensions_(dimensions) {
     Matrix<N, float> basis(1.0f);
-    for (int i = 0; i < N; ++i) {
+    for (unsigned int i = 0; i < N; ++i) {
         basis[i][i] = dimensions[i] * spacing[i];
     }
     
     this->setBasis(basis);
     Vector<N, float> offset(0.0f);
-    for (int i = 0; i < N; ++i) {
+    for (unsigned int i = 0; i < N; ++i) {
         offset += basis[i];
     }
     this->setOffset(-0.5f * offset);
@@ -362,12 +362,12 @@ void StructuredGridEntity<N>::setDimensions(const Vector<N, size_t>& dimensions)
 template <unsigned int N>
 Matrix<N + 1, float> StructuredGridEntity<N>::getIndexMatrix() const {
     Matrix<N + 1, float> indexMatrix(1.0f);
-    for (size_t i = 0; i < N; ++i) {
+    for (unsigned int i = 0; i < N; ++i) {
         indexMatrix[i][i] = static_cast<float>(dimensions_[i]);
     }
 
     // Offset to coordinates to center them in the middle of the texel/voxel.
-    for (size_t i = 0; i < N; i++) {
+    for (unsigned int i = 0; i < N; i++) {
         indexMatrix[N][i] = -0.5f;
     }
     return indexMatrix;
